+++ src/test/java/ampcontrol/model/training/model/layerblocks/graph/DummyOutputLayer.java
@@ -35,10 +35,10 @@ public class DummyOutputLayer implements LayerBlockConfig {
 
     /**
      * Sets weights output layer to the identity matrix.
-     * @param graph
+     * @param graph the {@link ComputationGraph} to modify
      */
     public static void setEyeOutput(ComputationGraph graph) {
-        final int[] wShape = graph.getOutputLayer(0).getParam("W").shape();
+        final long[] wShape = graph.getOutputLayer(0).getParam("W").shape();
         graph.getOutputLayer(0).setParam("W", Nd4j.eye(wShape[0]));
     }
 }
